requirements.txt
azure-core
azure-identity
azure-keyvault-secrets
python-dotenv
aiohttp
starlette==0.27.0
fastapi==0.100.0
uvicorn[standard]==0.22.0
pydantic==2.4.1
pydantic-settings==2.0.3
httpx==0.27.2
